backend/requirements.txt
flask==3.1.1
python-dotenv==1.1.0
flask-cors==5.0.0

langchain<1.0
langchain-community<1.0
langchain-openai<1.0
langchain-text-splitters<1.0

pypdf==5.6.1
sentence-transformers==4.1.0

chromadb>=0.5.0
langchain-chroma>=0.1.2
